DS-110-move expandable-content-directive under expandable-directives folder
