+++ ngx-fudis/projects/ngx-fudis/src/lib/components/expandable/expandable-directives/expandable-content.directive.ts
